requirements.txt
Flask>=1.0.2
Flask-SQLAlchemy>=2.3.2
pylast>=2.2.0
PyYAML>=3.12
uWSGI>=2.0.17
